requirements.txt
PyQt5
PyQtWebEngine
pillow
google-genai
together